requirements.txt
# 核心依赖
pycryptodome>=3.15.0
cryptography>=36.0.0
pywallet>=0.1.0

# 加密和哈希相关
pyaes==1.6.1
ecdsa==0.19.1
pbkdf2==1.3
sha256==1.0
pyOpenSSL==23.2.0

# 比特币相关
base58==0.2.2
mnemonic==0.13

# 工具和实用程序
tqdm>=4.65.0
requests>=2.11.1
PyYAML>=6.0.2
tabulate>=0.9.0

# 开发和测试依赖
pytest>=8.3.5
flake8>=7.1.2

# 可选依赖
# btcrecover>=0.11.0  # 支持更多钱包类型
# dbm  # Python内置模块，用于Berkeley DB支持
# sqlite3  # Python内置模块，用于SQLite支持

# 系统要求
# hashcat  # 系统需要安装hashcat用于GPU加速
# bitcoin-cli  # 系统需要安装Bitcoin Core用于直接与钱包交互

# 其他依赖
six>=1.17.0
python-dateutil>=2.9.0
